requirements.txt
streamlit
pandas
plotly
scipy
faker
google-generativeai
scikit-learn
joblib
